+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/custom.py
@@ -311,6 +311,96 @@ def k8s_install_cli(client_version='latest', install_location=None):
         raise CLIError('Connection error while attempting to download client ({})'.format(ex))
 
 
+def k8s_install_connector(client, name, resource_group, connector_name,
+                          location=None, service_principal=None, client_secret=None,
+                          chart_url=None, os_type='Linux'):
+    from subprocess import PIPE, Popen
+    helm_not_installed = "Helm not detected, please verify if it is installed."
+    image_tag = 'latest'
+    url_chart = chart_url
+    # Check if Helm is installed locally
+    try:
+        Popen(["helm"], stdout=PIPE, stderr=PIPE)
+    except OSError:
+        raise CLIError(helm_not_installed)
+    # Validate if the RG exists
+    groups = _resource_client_factory().resource_groups
+    # Just do the get, we don't need the result, it will error out if the group doesn't exist.
+    rgkaci = groups.get(resource_group)
+    # Auto assign the location
+    if location is None:
+        location = rgkaci.location  # pylint:disable=no-member
+    # Get the credentials from a AKS instance
+    _, browse_path = tempfile.mkstemp()
+    aks_get_credentials(client, resource_group, name, admin=False, path=browse_path)
+    subscription_id = _get_subscription_id()
+    dns_name_prefix = _get_default_dns_prefix(connector_name, resource_group, subscription_id)
+    # Ensure that the SPN exists
+    principal_obj = _ensure_service_principal(service_principal, client_secret, subscription_id,
+                                              dns_name_prefix, location, connector_name)
+    client_secret = principal_obj.get("client_secret")
+    service_principal = principal_obj.get("service_principal")
+    # Get the TenantID
+    profile = Profile()
+    _, _, tenant_id = profile.get_login_credentials()
+    # Check if we want the windows connector
+    if os_type.lower() in ['windows', 'both']:
+        # The current connector will deploy two connectors, one for Windows and another one for Linux
+        image_tag = 'canary'
+    logger.warning('Deploying the aci-connector using Helm')
+    try:
+        subprocess.call(["helm", "install", url_chart, "--name", connector_name, "--set", "env.azureClientId=" +
+                         service_principal + ",env.azureClientKey=" + client_secret + ",env.azureSubscriptionId=" +
+                         subscription_id + ",env.azureTenantId=" + tenant_id + ",env.aciResourceGroup=" + rgkaci.name +
+                         ",env.aciRegion=" + location + ",image.tag=" + image_tag])
+    except subprocess.CalledProcessError as err:
+        raise CLIError('Could not deploy the ACI Chart: {}'.format(err))
+
+
+def k8s_uninstall_connector(client, name, connector_name, resource_group,
+                            graceful=False, os_type='Linux'):
+    from subprocess import PIPE, Popen
+    helm_not_installed = "Error : Helm not detected, please verify if it is installed."
+    # Check if Helm is installed locally
+    try:
+        Popen(["helm"], stdout=PIPE, stderr=PIPE)
+    except OSError:
+        raise CLIError(helm_not_installed)
+    # Get the credentials from a AKS instance
+    _, browse_path = tempfile.mkstemp()
+    aks_get_credentials(client, resource_group, name, admin=False, path=browse_path)
+    if graceful:
+        logger.warning('Graceful option selected, will try to drain the node first')
+        kubectl_not_installed = "Kubectl not detected, please verify if it is installed."
+        try:
+            Popen(["kubectl"], stdout=PIPE, stderr=PIPE)
+        except OSError:
+            raise CLIError(kubectl_not_installed)
+        try:
+            if os_type.lower() in ['windows', 'both']:
+                nodes = ['-0', '-1']
+                for n in nodes:
+                    drain_node = subprocess.check_output(
+                        ["kubectl", "drain", "aci-connector{}".format(n), "--force"],
+                        universal_newlines=True)
+            else:
+                drain_node = subprocess.check_output(
+                    ["kubectl", "drain", "aci-connector", "--force"],
+                    universal_newlines=True)
+        except subprocess.CalledProcessError as err:
+            raise CLIError('Could not find the node, make sure you' +
+                           ' are using the correct --os-type option: {}'.format(err))
+        if not drain_node:
+            raise CLIError('Could not find the node, make sure you' +
+                           ' are using the correct --os-type')
+
+    logger.warning('Undeploying the aci-connector using Helm')
+    try:
+        subprocess.call(["helm", "del", connector_name, "--purge"])
+    except subprocess.CalledProcessError as err:
+        raise CLIError('Could not deploy the ACI Chart: {}'.format(err))
+
+
 def _build_service_principal(client, name, url, client_secret):
     # use get_progress_controller
     hook = APPLICATION.get_progress_controller(True)
@@ -960,6 +1050,7 @@ def create_application(client, display_name, homepage, identifier_uris,
                        available_to_other_tenants=False, password=None, reply_urls=None,
                        key_value=None, key_type=None, key_usage=None, start_date=None,
                        end_date=None):
+    from azure.graphrbac.models import GraphErrorException
     password_creds, key_creds = _build_application_creds(password, key_value, key_type,
                                                          key_usage, start_date, end_date)
 
@@ -970,7 +1061,14 @@ def create_application(client, display_name, homepage, identifier_uris,
                                                    reply_urls=reply_urls,
                                                    key_credentials=key_creds,
                                                    password_credentials=password_creds)
-    return client.create(app_create_param)
+    try:
+        return client.create(app_create_param)
+    except GraphErrorException as ex:
+        if 'insufficient privileges' in str(ex).lower():
+            link = 'https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-create-service-principal-portal'  # pylint: disable=line-too-long
+            raise CLIError("Directory permission is needed for the current user to register the application. "
+                           "For how to configure, please refer '{}'. Original error: {}".format(link, ex))
+        raise
 
 
 def _build_application_creds(password=None, key_value=None, key_type=None,
@@ -1373,7 +1471,7 @@ def _print_or_merge_credentials(path, kubeconfig):
 
     # ensure that at least an empty ~/.kube/config exists
     directory = os.path.dirname(path)
-    if not os.path.exists(directory):
+    if directory and not os.path.exists(directory):
         try:
             os.makedirs(directory)
         except OSError as ex: